moving generic classes
