+++ client-src/ui/components/Editor/prosemirror/EditorNode.types.ts
@@ -1,5 +1,5 @@
 import { NodeSpec } from "prosemirror-model";
-import { type } from "../../../../../type/inject";
+import { type } from "../../../../type/inject";
 
 export const EditorNode = type<EditorNode>();
 export interface EditorNode {
