+++ client-src/ui/components/Editor/prosemirror/EditorSchema.ts
@@ -1,7 +1,7 @@
 import { Schema, NodeSpec, MarkSpec } from "prosemirror-model";
 import { inject, dependency } from "../../../../type/inject";
-import { EditorNode } from "./nodes/EditorNode.types";
-import { EditorMark } from "./marks/EditorMark.types";
+import { EditorNode } from "./EditorNode.types";
+import { EditorMark } from "./EditorMark.types";
 import { EditorSchema as IEditorSchema } from "./EditorSchema.types";
 
 @dependency(IEditorSchema)
