+++ client-src/ui/components/Editor/prosemirror/nodes/HeadingNode.ts
@@ -1,4 +1,4 @@
-import { EditorNode } from "./EditorNode.types";
+import { EditorNode } from "../EditorNode.types";
 import { dependency } from "../../../../../type/inject";
 import { NodeSpec } from "prosemirror-model";
